linux: get mcpp to compile and link on Linux
Fixes:
- ARGB.h: cast 0LL/255LL to int64_t to disambiguate clamp() overload
  (Linux gcc has int64_t=long, not long long, causing ambiguity)
- Minecraft.cpp: clean up botched loadResourceTex/loadResourceText
  duplicates; add Linux branches that return empty
- Minecraft.cpp: guard SoundEngine::setListenerPosition calls
  behind #ifdef _WIN32; use SoundManager no-op on Linux
- SoundManager.h: include OggDecoder.h for OggData type; add
  setListenerPosition() no-op stub; guard SoundEngine.h include
- TitleScreen.cpp: use std::exit(0) on Linux for Quit button
- Connection.cpp: include <netinet/tcp.h> for TCP_NODELAY,
  add full POSIX socket headers
- LevelRenderer.cpp: load atlas from AssetManager on Linux
  (no Win32 RCDATA resources available)
- Window_glfw.cpp: remove duplicate nativeHandle()/hwnd() defs
  (already inline in Window.h)
- RenderBackend.cpp: use unified DeviceGL (already cross-platform),
  drop broken DeviceGL_glfw.h shim
- CMakeLists.txt: skip asset_packer on Linux (no Windows resources);
  use find_library for libGL (system preferred, /tmp/x11-dev fallback);
  link OpenSSL + Threads on Linux
- src/CMakeLists.txt: skip debug_test target on Linux (WinMain-only)

Result: mcpp builds cleanly on Linux x86_64 (61 MB ELF executable).
Parity test binaries also build. The binary launches and reaches
asset loading before failing (expected: assets.bin not yet built
for Linux — needs asset_packer port or runtime asset dir setup).

+++ src/util/ARGB.h
@@ -94,9 +94,9 @@ inline int scaleRGB(int color_, float sR, float sG, float sB) {
 inline int scaleRGB(int color_, float scale) { return scaleRGB(color_, scale, scale, scale); }
 inline int scaleRGB(int color_, int scale) {
     return color(alpha(color_),
-                 static_cast<int>(mth::clamp(static_cast<int64_t>(red(color_)) * scale / 255LL, 0LL, 255LL)),
-                 static_cast<int>(mth::clamp(static_cast<int64_t>(green(color_)) * scale / 255LL, 0LL, 255LL)),
-                 static_cast<int>(mth::clamp(static_cast<int64_t>(blue(color_)) * scale / 255LL, 0LL, 255LL)));
+                 static_cast<int>(mth::clamp(static_cast<int64_t>(red(color_)) * scale / 255LL, static_cast<int64_t>(0), static_cast<int64_t>(255))),
+                 static_cast<int>(mth::clamp(static_cast<int64_t>(green(color_)) * scale / 255LL, static_cast<int64_t>(0), static_cast<int64_t>(255))),
+                 static_cast<int>(mth::clamp(static_cast<int64_t>(blue(color_)) * scale / 255LL, static_cast<int64_t>(0), static_cast<int64_t>(255))));
 }
 inline int greyscale(int color_) {
     int g = static_cast<int>(red(color_) * 0.3F + green(color_) * 0.59F + blue(color_) * 0.11F);
